--- libs/service-portal/family/src/screens/FamilyOverview/FamilyOverview.css.ts
@@ -1,6 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-export const avatar = style({
-  width: 80,
-  height: 80,
-})
