--- libs/service-portal/licenses/src/components/LicenseIcon/LicenseIcon.css.ts
@@ -1,4 +1,0 @@
-import { style } from '@vanilla-extract/css'
-import { theme } from '@island.is/island-ui/theme'
-
-export const logo = style({})
