+++ libs/service-portal/licenses/src/screens/DrivingLicenseDetail/DrivingLicenseDetail.tsx
@@ -24,7 +24,7 @@ import { isExpired, toDate } from '../../utils/dateUtils'
 import { mapCategory } from '../../utils/dataMapper'
 import ReactHtmlParser from 'react-html-parser'
 import ExpandableLine from './ExpandableLine'
-import * as styles from '../../components/DrivingLicense/DrivingLicense.css'
+import * as styles from './DrivingLicenseDetail.css'
 import QRCodeModal from '../../components/QRCodeModal/QRCodeModal'
 import { info } from 'kennitala'
 import { m } from '../../lib/messages'
